requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
django-multiselectfield==0.1.12
djangorestframework==3.12.1
freeze==3.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
